+++ backend/app/chatgpt.py
@@ -61,31 +61,6 @@ my_functions = [
 ]
 
 
-def create_event(
-    summary,
-    start,
-    end,
-    location=None,
-    description=None,
-    member=None,
-    timeZone="Asia/Tokyo",
-):
-    event = {
-        "summary": summary,
-        "location": location,
-        "description": None if member is None else f"参加者: {member}",
-        "start": {
-            "dateTime": start,
-            "timeZone": timeZone,
-        },
-        "end": {
-            "dateTime": end,
-            "timeZone": timeZone,
-        },
-    }
-    return event
-
-
 def send_chat(
     user_message: str, system_message: str = DEFAULT_SYSTEM_PROMPT
 ) -> openai.ChatCompletion:
@@ -102,29 +77,7 @@ def send_chat(
     print("prompt:", system_message, user_message)
     print("Got response:", res)
 
-    add_event_if_triggered(res)
-
     return res
-
-
-def add_event_if_triggered(chat_response: dict):
-    if not is_function_triggered(chat_response):
-        return False
-
-    # 関数を使用すると判断された場合
-
-    message = chat_response["choices"][0]["message"]
-    # 使うと判断された関数名
-    function_name = message["function_call"]["name"]
-    # その時の引数dict
-    arguments = json.loads(message["function_call"]["arguments"])
-
-    print(function_name)
-    print(arguments)
-
-    event = create_event(**arguments)
-
-    add_event_to_calendar(event)
 
 
 def is_function_triggered(chat_response: dict):
